feat: Enhance post update and delete functionality with user ownership checks

+++ server-side/src/modules/posts/posts.service.ts
@@ -3,6 +3,7 @@ import {
   Injectable,
   NotFoundException,
   Logger,
+  ForbiddenException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Post, PostDocument } from './post.schema';
@@ -44,8 +45,13 @@ export class PostsService {
     return post;
   }
 
-  async update(id: string, updatePostDto: UpdatePostDto): Promise<Post> {
+  async update(
+    id: string,
+    updatePostDto: UpdatePostDto,
+    userId: string,
+  ): Promise<Post> {
     validateId(id);
+    await this.checkOwnership(id, userId);
     const updatedPost = await this.postModel
       .findByIdAndUpdate(id, updatePostDto, {
         new: true,
@@ -60,8 +66,9 @@ export class PostsService {
     return updatedPost;
   }
 
-  async delete(id: string): Promise<string> {
+  async delete(id: string, userId: string): Promise<string> {
     validateId(id);
+    await this.checkOwnership(id, userId);
     this.logger.log(`Deleting post with ID: ${id}`);
     const result = await this.postModel.findByIdAndDelete(id).lean().exec();
     if (!result) {
@@ -74,5 +81,17 @@ export class PostsService {
 
   async count(): Promise<number> {
     return this.postModel.countDocuments();
+  }
+  async checkOwnership(postId: string, userId: string): Promise<Post> {
+    const post = await this.postModel.findById(postId).exec();
+    if (!post) {
+      throw new NotFoundException(`Post with ID ${postId} not found`);
+    }
+    if (post.user.toString() !== userId) {
+      throw new ForbiddenException(
+        'You are not authorized to modify this post',
+      );
+    }
+    return post;
   }
 }
